backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-socketio==5.10.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
aiofiles==23.2.1
reportlab==4.0.7
email-validator==2.1.0
